+++ Assets/Scripts/Dialogue/DialogueService.cs
@@ -3,7 +3,7 @@
 public class DialogueService
 {
     private DialogueRepository dialogueRepository = new DialogueRepository();
-    public DialogueObject GetDialogueObject(float id)
+    public DialogueObject GetDialogueObject(int id)
     {
         DialogueObject dialogueObject = dialogueRepository.ReadDialogueObjectById(id);
         foreach (int lineId in dialogueObject.dialogueLineIds)
